fix: binary sensors were added as regular sensors refactor: code cleanup
