+++ custom_components/ina219_ups_hat/entity.py
@@ -1,13 +1,18 @@
 
 from .const import DOMAIN
 from .coordinator import INA219UpsHatCoordinator
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
+from homeassistant.helpers.device_registry import DeviceInfo
 
 
 class INA219UpsHatEntity():
     def __init__(self, coordinator: INA219UpsHatCoordinator) -> None:
         self._coordinator = coordinator
         self._device_id = self._coordinator.id_prefix
+        self._attr_device_info = DeviceInfo(
+            identifiers={(DOMAIN, coordinator.id_prefix)},
+            name=coordinator.name_prefix,
+            manufacturer="Some Chinese factory",
+        )
 
     @property
     def name(self):
@@ -16,16 +21,6 @@ class INA219UpsHatEntity():
     @property
     def unique_id(self):
         return self._coordinator.id_prefix + '_' + self._name
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device_info of the device."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._coordinator.id_prefix)},
-            name=self._coordinator.name_prefix,
-            manufacturer="Some Chinese factory",
-            sw_version="0.3.0",
-        )
 
     async def async_update(self):
         await self._coordinator.async_request_refresh()
