+++ custom_components/ina219_ups_hat/manifest.json
@@ -8,5 +8,5 @@
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/odya/hass-ina219-ups-hat/issues",
   "requirements": ["smbus2>=0.4.2","numpy>=1.23.2"],
-  "version": "0.3.0"
+  "version": "0.3.1"
 }
